move gitb from db to misc
